darkweb-threat-intel/backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
httpx
openai
tenacity
pandas
torch
scikit-learn
numpy
fastapi
uvicorn
httpx
